move contractInteraction logic to own component
