+++ packages/nextjs/components/stake/index.tsx
@@ -1,1 +1,2 @@
 export * from "./EthToPrice";
+export * from "./StakeContractInteraction";
